# src/main/java/com/katherina/nuevospa/taskmanagerapi/dto/ErrorResponse.java
package com.katherina.nuevospa.taskmanagerapi.dto;

import java.time.LocalDateTime;

public record ErrorResponse(
        LocalDateTime timestamp,
        int status,
        String error,
        String message,
        String path
) {
    public ErrorResponse(int status, String error, String message, String path) {
        this(LocalDateTime.now(), status, error, message, path);
    }
}