# src/main/resources/openapi/openapi.yml
openapi: 3.0.3
info:
  title: Task Management API - NUEVO SPA
  description: API RESTful para gestión de tareas con autenticación JWT
  version: 1.0.0
  contact:
    name: NUEVO SPA
servers:
  - url: http://localhost:8080/api
    description: Local development server

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token de autenticación

  schemas:
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: admin@nuevospa.com
        password:
          type: string
          format: password
          example: admin123

    LoginResponse:
      type: object
      properties:
        token:
          type: string
          description: JWT access token
        tokenType:
          type: string
          example: Bearer
        email:
          type: string
          format: email

    TaskRequest:
      type: object
      required:
        - title
        - description
        - taskStatusId
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
          example: Implementar módulo de reportes
        description:
          type: string
          maxLength: 1000
          example: Crear los reportes mensuales del sistema
        taskStatusId:
          type: integer
          format: int64
          example: 1

    TaskResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
        createDate:
          type: string
          format: date-time
        updateDate:
          type: string
          format: date-time
        taskStatus:
          $ref: '#/components/schemas/TaskStatusResponse'
        user:
          $ref: '#/components/schemas/UserResponse'

    TaskStatusResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          example: PENDIENTE

    UserResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        email:
          type: string
          format: email

    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string

paths:
  /auth/login:
    post:
      tags:
        - Autenticación
      summary: Iniciar sesión
      description: Autentica un usuario y retorna un token JWT
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Autenticación exitosa
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Credenciales inválidas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /tasks:
    get:
      tags:
        - Tareas
      summary: Listar todas las tareas
      description: Obtiene la lista de todas las tareas del usuario autenticado
      operationId: listarTareas
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de tareas obtenida exitosamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TaskResponse'
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Tareas
      summary: Crear una tarea
      description: Crea una nueva tarea para el usuario autenticado
      operationId: createTask
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskRequest'
      responses:
        '201':
          description: Tarea creada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          description: Datos de entrada inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /tareas/{id}:
    get:
      tags:
        - Tareas
      summary: Obtener una tarea por ID
      description: Obtiene los detalles de una tarea específica
      operationId: getTask
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Tarea obtenida exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '404':
          description: Tarea no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Tareas
      summary: Actualizar una tarea
      description: Actualiza los datos de una tarea existente
      operationId: updateTask
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskRequest'
      responses:
        '200':
          description: Tarea actualizada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          description: Datos de entrada inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Tarea no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Tareas
      summary: Eliminar una tarea
      description: Elimina una tarea existente
      operationId: deleteTask
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Tarea eliminada exitosamente
        '404':
          description: Tarea no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'